feat: merge core and compiler
