+++ cli/src/query_repl.rs
@@ -14,9 +14,9 @@ use rustyline::validate::MatchingBracketValidator;
 use rustyline::{Cmd, CompletionType, Config, EditMode, Editor, KeyEvent};
 use rustyline::{Helper, Validator};
 
-use typst_ts_compiler::{CompileDriver, CompileReport, ConsoleDiagReporter, PureCompiler};
-use typst_ts_compiler::{ShadowApiExt, TypstSystemWorld};
 use typst_ts_core::{typst::prelude::*, GenericExporter};
+use typst_ts_core::{CompileDriver, CompileReport, ConsoleDiagReporter, PureCompiler};
+use typst_ts_core::{ShadowApiExt, TypstSystemWorld};
 
 use crate::query::serialize;
 use crate::CompileOnceArgs;
