+++ crates/reflexo-typst/src/features.rs
@@ -1,7 +1,7 @@
 use once_cell::sync::Lazy;
-use typst_ts_core::typst::prelude::*;
 
 use super::diag::DiagnosticFormat;
+use crate::typst::prelude::*;
 
 #[derive(Debug, Clone, Copy)]
 pub struct FeatureSlot(u16);
