+++ fuzzers/incremental/src/main.rs
@@ -2,10 +2,6 @@ use std::path::Path;
 
 use reflexo_typst2vec::incr::{IncrDocClient, IncrDocServer};
 use typst::model::Document;
-use typst_ts_compiler::{
-    CompileDriver, CompileExporter, PureCompiler, ShadowApiExt, TypstSystemUniverse,
-    TypstSystemWorld,
-};
 use typst_ts_core::{
     config::{entry::EntryOpts, CompileOpts},
     exporter_builtins::GroupExporter,
@@ -13,6 +9,10 @@ use typst_ts_core::{
         ir::{Abs, Point, Rect},
         stream::BytesModuleStream,
     },
+};
+use typst_ts_core::{
+    CompileDriver, CompileExporter, PureCompiler, ShadowApiExt, TypstSystemUniverse,
+    TypstSystemWorld,
 };
 use typst_ts_incremental_fuzzer::mutate;
 use typst_ts_svg_exporter::IncrSvgDocClient;
