+++ packages/parser/src/lib.rs
@@ -2,10 +2,10 @@ use std::{path::Path, sync::Arc};
 
 use js_sys::Uint32Array;
 use typst::syntax::{FileId, VirtualPath};
-use typst_ts_compiler::parser::{
+use typst_ts_core::error::prelude::*;
+use typst_ts_core::parser::{
     get_semantic_tokens_full, get_semantic_tokens_legend, OffsetEncoding, SemanticToken,
 };
-use typst_ts_core::error::prelude::*;
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
