+++ tests/integration/src/lib.rs
@@ -2,16 +2,14 @@ pub mod wasm;
 
 use std::path::Path;
 
-use typst_ts_compiler::{
-    CompileDriver, CompileExporter, PureCompiler, TypstSystemUniverse, TypstSystemWorld,
-};
 use typst_ts_core::PdfDocExporter;
 use typst_ts_core::SvgModuleExporter;
 use typst_ts_core::{
     config::{entry::EntryOpts, CompileOpts},
     exporter_builtins::{FsPathExporter, GroupExporter},
     path::PathClean,
-    TypstDocument,
+    CompileDriver, CompileExporter, PureCompiler, TypstDocument, TypstSystemUniverse,
+    TypstSystemWorld,
 };
 
 fn get_driver(
